Add files via upload
Part 1 of the **EM_Digit Processing Project** focuses on integrating, cleaning, and standardizing multiple datasets to create a structured and optimized data foundation for further analysis. This phase begins by importing key datasets, including geolocated information, full test data, and a gazetteer, and merging them efficiently to ensure accurate and up-to-date location references. The script prioritizes non-null values, updates missing fields, removes duplicates, and sorts the data for consistency. Advanced optimizations, such as vectorized operations and efficient merging strategies, significantly improve performance and memory usage. Additionally, the processed data is exported and uploaded to Google Sheets for accessibility and collaboration. By the end of this phase, the dataset is fully prepared for further enrichment, validation, and geographical processing in the next stage of the project.